+++ packages/cli/babel-preset-app/CHANGELOG.md
@@ -87,7 +87,7 @@
 
 ### Minor Changes
 
-- 7b9067f: add babel plugin for webpack-builder
+- 7b9067f: add babel plugin for builder
 
 ### Patch Changes